= SipClientMFC =  * 64bit 컴파일 옵션을 수정함

+++ trunk/SipUserAgent/SipUserAgentCall.hpp
@@ -169,7 +169,7 @@ int CSipUserAgent::GetCallCount( )
 	int iCallCount;
 
 	m_clsMutex.acquire();
-	iCallCount = m_clsMap.size();
+	iCallCount = (int)m_clsMap.size();
 	m_clsMutex.release();
 
 	return iCallCount;
